requirements.txt
asn1crypto==0.24.0
aws-requests-auth==0.4.2
boto==2.48.0
boto3==1.7.48
botocore==1.10.48
cffi==1.11.5
cryptography==2.4.2
dataclasses<=0.6
elasticsearch>=5.0.0,<6.0.0
elasticsearch-dsl>=5.0.0,<6.0.0
jsonobject==0.7.1
Faker==0.8.11
future==0.16.0
hca==4.4.1
git+git://github.com/hannes-ucsc/jmespath.py@hannes-fix-cache-race#egg=jmespath
git+git://github.com/HumanCellAtlas/metadata-api@release/1.0b8#egg=hca-metadata-api
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.6.4
requests==2.20.0
requests-http-signature==0.0.3
rsa==4.0
urllib3==1.23
more-itertools==4.3.0
